# Faster conversion GAP <-> fmpz