This is a draft because we can not yet switch to the newer Nemo.jl required by this PR; see PR #218 for the state of that.

Suppose the development version of Oscar.jl was loaded and the following assignments were made:

    x = @gap 5^200;
    y = fmpz(5)^200;

Before this patch:

    julia> @btime GAP.gap_to_julia(fmpz, x);
      254.204 ns (4 allocations: 112 bytes)

    julia> @btime GAP.julia_to_gap(y);
      215.226 ns (5 allocations: 192 bytes)

After this patch:

    julia> @btime GAP.gap_to_julia(fmpz, x);
      83.054 ns (1 allocation: 16 bytes)

    julia> @btime GAP.julia_to_gap(y);
      41.068 ns (2 allocations: 96 bytes)

Resolves #202